fix: address PR #66 code review findings (critical + important)
- handler.rs: add parser mutex poison handling to trigger() and
  accept_with_chaining(), with two regression tests modeled on the
  existing render/accept_suggestion coverage (fixes C1)
- handler.rs: clear dynamic_task alongside dynamic_rx in stale/empty/
  Disconnected branches of try_merge_dynamic so dismiss()/trigger()
  aborts don't rely on already-completed JoinHandles (fixes I4)
- proxy.rs + gc-suggest: deduplicate resolve_spec_dirs and
  partition_spec_dirs into a single shared helper at
  gc-suggest/src/spec_dirs.rs; ghost-complete status/validate-specs
  now display the same spec dirs the proxy loads, with identical
  validation/fallback/warning semantics (fixes C2)
- proxy.rs: break Task B loop inside the CPR rollback poison branch
  instead of continuing with permanently elevated cpr_pending that
  would otherwise make Task A steal the next real application CPR
  response (fixes I3)
- proxy.rs: add INVARIANT comment explaining why the startup-only
  .expect() on handler mutex during Task E setup is currently safe,
  and warning future refactorers to convert it if the invariant
  breaks (fixes I6)
